scripts: Add a script to list changes since last lmp-manifest update
This script will list all "Relevant changes" in the meta-lmp
repository since the last lmp-manifest update.

It requires the meta-lmp repository checked out and current.

This assumes you are in the top directory of lmp-manifest repo
and the meta-lmp repo is checked out as ../meta-lmp.

It does however have options to specify where the meta-lmp repo
is checked out.

Also, it does not look in the checked out lmp-manifest but grabs
the last meta-lmp revision from github.com/foundriesio/lmp-manifest
directly.

Signed-off-by: Tim Anderson <tim.anderson@foundries.io>